[FD-429] refactor: TeamPlanOrchestrator 생성
- 도메인 로직과 애플리케이션 계층 로직 분리
